requirements.txt
jira
gitpython
Flask<1
keyring
python-dateutil